Remove `AttributeLintKind::UnknownCrateTypesSuggestion` variant
